Extract morphir.kyox.core and morphir.langkit.core from the Elm langkit
Two clusters had landed under morphir.langkit.elm only because the port
renamed krueger's top-level namespace wholesale; neither is Elm-specific.

morphir/kyox/core/ (package morphir.kyox.log) takes the Kyo Log <-> scribe
bridge: ScribeLogHandler, InMemoryLogRecorder, ScribeLogLayer. Home for
general Kyo extensions going forward. The Elm langkit was the only module
carrying a scribe dependency for it, so that dependency moves too.

morphir/langkit/core/ (package morphir.langkit.core) takes the pieces of
the diagnostic machinery with no Elm content: SourceOffsets, SourceSnippet
/SourceSnippetBuilder, DiagnosticContextLine, and SourceSpan.

Left in the Elm langkit, since each is Elm-specific in substance rather
than by accident: DiagnosticCode (its newtype validates ELM-P###/ELM-T###
codes), DiagnosticMessageFormatter (Elm-inspired prose), ParseDiagnostic
(Elm parse/tokenize factories), and CompileError.
