+++ morphir/kyox/core/src/morphir/kyox/log/InMemoryLogRecorder.scala
@@ -1,4 +1,4 @@
-package morphir.langkit.elm.log
+package morphir.kyox.log
 
 import kyo.*
