+++ morphir/langkit/core/test/src/morphir/langkit/core/SourceSnippetBuilderSpec.scala
@@ -1,4 +1,4 @@
-package morphir.langkit.elm.compiler
+package morphir.langkit.core
 
 import kyo.test.*
